+++ include/rocksdb/c.h
@@ -899,16 +899,16 @@ rocksdb_options_set_min_indirect_val_size(
     rocksdb_options_t*, uint64_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_fraction_remapped_during_compaction(
-    rocksdb_options_t*, double* values, size_t num_levels);
+    rocksdb_options_t*, int32_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_fraction_remapped_during_active_recycling(
-    rocksdb_options_t*, double* values, size_t num_levels);
+    rocksdb_options_t*, int32_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_fragmentation_active_recycling_trigger(
-    rocksdb_options_t*, double* values, size_t num_levels);
+    rocksdb_options_t*, int32_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_fragmentation_active_recycling_klaxon(
-    rocksdb_options_t*, double* values, size_t num_levels);
+    rocksdb_options_t*, int32_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_active_recycling_sst_minct(
     rocksdb_options_t*, int32_t* values, size_t num_levels);
@@ -923,7 +923,7 @@ rocksdb_options_set_vlogfile_max_size(
     rocksdb_options_t*, uint64_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_compaction_picker_age_importance(
-    rocksdb_options_t*, double* values, size_t num_levels);
+    rocksdb_options_t*, int32_t* values, size_t num_levels);
 extern ROCKSDB_LIBRARY_API void
 rocksdb_options_set_ring_compression_style(
     rocksdb_options_t*, int32_t* values, size_t num_levels);
