+++ options/cf_options.h
@@ -203,15 +203,15 @@ struct MutableCFOptions {
         compaction_score_limit_L0(0),
         vlogring_activation_level(std::vector<int32_t>({0})),
         min_indirect_val_size(std::vector<uint64_t>({0})),
-        fraction_remapped_during_compaction(std::vector<double>({0})),
-        fraction_remapped_during_active_recycling(std::vector<double>({0})),
-        fragmentation_active_recycling_trigger(std::vector<double>({0})),
-        fragmentation_active_recycling_klaxon(std::vector<double>({0})),
+        fraction_remapped_during_compaction(std::vector<int32_t>({0})),
+        fraction_remapped_during_active_recycling(std::vector<int32_t>({0})),
+        fragmentation_active_recycling_trigger(std::vector<int32_t>({0})),
+        fragmentation_active_recycling_klaxon(std::vector<int32_t>({0})),
         active_recycling_sst_minct(std::vector<int32_t>({0})),
         active_recycling_sst_maxct(std::vector<int32_t>({0})),
         active_recycling_vlogfile_freed_min(std::vector<int32_t>({0})),
         vlogfile_max_size(std::vector<uint64_t>({0})),
-        compaction_picker_age_importance(std::vector<double>({0})),
+        compaction_picker_age_importance(std::vector<int32_t>({0})),
     //Ring Compression Style: indicates what kind of compression will be applied to the data
         ring_compression_style(std::vector<CompressionType>({kNoCompression})),
 #endif
@@ -269,15 +269,15 @@ struct MutableCFOptions {
 // options related to ring structure and compression are immutable.  Ring structure may change over a restart, as long as a ring is not deleted while it is holding values
   std::vector<int32_t> vlogring_activation_level;
   std::vector<uint64_t> min_indirect_val_size;
-  std::vector<double> fraction_remapped_during_compaction;
-  std::vector<double> fraction_remapped_during_active_recycling;
-  std::vector<double> fragmentation_active_recycling_trigger;
-  std::vector<double> fragmentation_active_recycling_klaxon;
+  std::vector<int32_t> fraction_remapped_during_compaction;
+  std::vector<int32_t> fraction_remapped_during_active_recycling;
+  std::vector<int32_t> fragmentation_active_recycling_trigger;
+  std::vector<int32_t> fragmentation_active_recycling_klaxon;
   std::vector<int32_t> active_recycling_sst_minct;
   std::vector<int32_t> active_recycling_sst_maxct;
   std::vector<int32_t> active_recycling_vlogfile_freed_min;
   std::vector<uint64_t> vlogfile_max_size;
-  std::vector<double> compaction_picker_age_importance;
+  std::vector<int32_t> compaction_picker_age_importance;
   std::vector<CompressionType> ring_compression_style;
 #endif
